implement posts_helper spec

+++ app/helpers/posts_helper.rb
@@ -11,8 +11,10 @@ module PostsHelper
     nav_tag(
       content_tag(:ul) do
         concat content_tag(:li, link_to("Post list", forum_posts_path(@forum)))
-        concat content_tag(:li, link_to("Edit post", edit_forum_post_path(@forum, @post)))
-        concat content_tag(:li, link_to("Destroy post", forum_post_path(@forum, @post), :method => :delete))
+        if @post.editable_by?(current_user)
+          concat content_tag(:li, link_to("Edit post", edit_forum_post_path(@forum, @post)))
+          concat content_tag(:li, link_to("Destroy post", forum_post_path(@forum, @post), :method => :delete))
+        end
       end
     )
   end
